[#1539] Implement Clear button on events page
This commit implements a Clear button on the Events page.  Clicking the
button pops up a dialog asking to confirm whether to really delete
everything (since it can't be undone).  Confirming the dialog sends a
DELETE request to `/events` to instruct the backend to delete all of the
events.

The Clear button on the Events page is styled with a red outline and
text to indicate that it leads to a dangerous operation.  The Delete
button in the dialog is styled with a red background to indicate that it
actually does the dangerous option.

The dialog has an exclamation mark icon because the description text
always has a 0.5em left margin.  Rather than replace a whole section of
the dialog using templates, I put an relevant icon in.

The request to the backend uses DELETE on the collection endpoint, as I
believe is conventional for REST+JSON APIs.  The response for a
successful deletion is 204 No Content (because it's all gone).  Rather
than adding another round-trip to refresh the list, I replace the events
with an empty array on a successful response.  On error, the API returns
a conventional error message, which I display in a right-corner toast
notification.