Add more benchmark implementations, use Protobuf
After a lot of investigation, I think I've found the best way to
accomplish this bulk insert task.

The previous implementation used `COPY ... WITH CSV HEADER`.  While this
is the fastest possible way to accomplish the goal, it only works if the
CSV file can be copied to the host on which PostgreSQL is running.
Enabling this would require significant changes to the deployment
methods for Stork (the server host would need to be able to copy files
to the database host).  This amount of change is likely not acceptable
to customers.

Instead, I've put the data into a series of Protobuf messages (each
preceded by their length as an unsigned 16-bit integer in big-endian
order) and inserted it into the database using a prepared statement
running in a transaction.  This version runs 3% faster and uses 33% less
memory than an equivalent implementation using Go's CSV parser.
Unfortunately, the Protobuf/prepared statement implementation is, on
average, 275% slower and allocates 2,460% more memory than the `COPY ...
WITH CSV HEADER` implementation.  It loads 7 million records in 122
seconds.